requirements.txt
fastapi
uvicorn[standard]
websockets
google-cloud-speech
google-cloud-texttospeech
google-generativeai
openai
python-dotenv
aiofiles
pydantic-settings